levels/level_02_prompt_engineering/pyproject.toml
[project]
name = "level-02-prompt-engineering"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50",
    "pydantic>=2.9",
    "pyyaml>=6.0",
    "reflex>=0.6",
    "psutil>=6.0",
    "pynvml>=11.0",
    "requests>=2.31",
]
